Abstract error rendering to ComparisonTable
